+++ src/wxUI/menus.py
@@ -101,3 +101,14 @@ class attachMenu(wx.Menu):
 		self.Append(self.upload)
 		self.add = wx.MenuItem(self, wx.NewId(), _("Add from VK"))
 		self.Append(self.add)
+
+class communityBufferMenu(wx.Menu):
+	def __init__(self):
+		super(communityBufferMenu, self).__init__()
+		load = wx.Menu()
+		self.load_posts = load.Append(wx.NewId(), _("Load posts"))
+		self.load_topics = load.Append(wx.NewId(), _("Load topics"))
+		self.load_topics.Enable(False)
+		self.load_audios = load.Append(wx.NewId(), _("Load audios"))
+		self.load_videos = load.Append(wx.NewId(), _("Load videos"))
+		self.Append(wx.NewId(), _("Load"), load)