+++ changelog.md
@@ -23,6 +23,9 @@
     * Fixed errors when moving songs to albums. Now everything works as expected.
 * Added documents to the list of supported files when adding attachments to a wall post or private message.
 * It is possible to enable or disable proxy from the preferences dialog. The application must be restarted for this change to take effect.
+* Added improvements to groups:
+    * It is possible to load audios and videos for a group. In order to do so, you need to go to the group buffer and press the menu key, or right mouse click, in the tree item representing the group. New buffers will be created inside the current group's buffer.
+    * You can create or delete all buffers for groups by pressing the menu key or right mouse click in the "communities" buffer.
 
 ## Changes in version 0.18 (21.01.2019)